Reordered some functions; added per beard help.
